Improves support for the attribute encoding strategy.
The attribute encoding strategy was only applied for built-in value types (String, Int, Float, Double, etc.) and for Date/NSDate, but not for properties that are naturally encoded as strings or numbers, in particular string or int enums. With this change, if a type includes e.g. a property `var choice: Choice` where `Choice` is an `enum Choice: String, Encodable`, then the property `choice` will be included as part of the `attributeEncodingStrategy` of the XML encoder.